Convert sections to classmethods
