+++ great_expectations/render/view_models/default/section/__init__.py
@@ -1,0 +1,2 @@
+from .prescriptive import PrescriptiveExpectationColumnSectionRenderer
+from .descriptive import DescriptiveEvrColumnSectionRenderer
